requirements.txt
markdown==3.5.2
Jinja2==3.1.6
PyYAML==6.0.1
Flask==3.0.2
Werkzeug==3.1.3
weasyprint==60.2
requests==2.31.0
huggingface-hub==0.23.0 